small fixes after personal review
