--- src/VirtoCommerce.PageBuilderModule.Core/Models/FilePublishStatus.cs
@@ -1,8 +1,0 @@
-//namespace VirtoCommerce.PageBuilderModule.Core.Models;
-
-//public class FilePublishStatus
-//{
-//    public string Status { get; set; }
-//    public bool Published { get; set; }
-//    public bool HasChanges { get; set; }
-//}
