+++ src/VirtoCommerce.PageBuilderModule.Core/ModuleConstants.cs
@@ -35,7 +35,7 @@ namespace VirtoCommerce.PageBuilderModule.Core
                     Read,
                     Update,
                     Delete,
-                    Publish
+                    Publish,
                 ];
             }
         }
