+++ src/VirtoCommerce.PageBuilderModule.Core/Services/IPageBuilderPageService.cs
@@ -3,7 +3,4 @@ using VirtoCommerce.Platform.Core.GenericCrud;
 
 namespace VirtoCommerce.PageBuilderModule.Core.Services;
 
-public interface IPageBuilderPageService : ICrudService<PageBuilderPage>
-{
-    Task<string> GetPageContentAsync(string pageId, CancellationToken cancellationToken = default);
-}
+public interface IPageBuilderPageService : ICrudService<PageBuilderPage>;
