+++ src/VirtoCommerce.PageBuilderModule.Data/Services/PageBuilderPageService.cs
@@ -1,4 +1,3 @@
-using System.Text;
 using VirtoCommerce.PageBuilderModule.Core.Events;
 using VirtoCommerce.PageBuilderModule.Core.Models;
 using VirtoCommerce.PageBuilderModule.Core.Services;
@@ -13,7 +12,6 @@ namespace VirtoCommerce.PageBuilderModule.Data.Services;
 
 public class PageBuilderPageService(
     Func<IPageBuilderModuleRepository> repositoryFactory,
-    Func<IContentStreamRepository> contentStreamRepositoryFactory,
     IPlatformMemoryCache platformMemoryCache,
     IEventPublisher eventPublisher)
     : CrudService<PageBuilderPage, PageBuilderPageEntity, PageBuilderPageChangingEvent, PageBuilderPageChangedEvent>(
@@ -22,17 +20,5 @@ public class PageBuilderPageService(
     protected override Task<IList<PageBuilderPageEntity>> LoadEntities(IRepository repository, IList<string> ids, string responseGroup)
     {
         return ((IPageBuilderModuleRepository)repository).GetPageBuilderPagesByIdsAsync(ids, responseGroup);
-    }
-
-    public async Task<string> GetPageContentAsync(string pageId, CancellationToken cancellationToken = default)
-    {
-        var repository = contentStreamRepositoryFactory();
-        using var stream = new MemoryStream();
-        await using var writer = new StreamWriter(stream, Encoding.UTF8, bufferSize: ContentStreamRepository.ContentBufferSize, leaveOpen: true);
-        await repository.LoadBinaryAsync(pageId, writer, cancellationToken);
-        await writer.FlushAsync(cancellationToken);
-        stream.Position = 0;
-        using var reader = new StreamReader(stream, Encoding.UTF8, detectEncodingFromByteOrderMarks: true, leaveOpen: false);
-        return await reader.ReadToEndAsync(cancellationToken);
     }
 }
